Cleanup SPIFloat and default values in styles
Remove remaining uses of SCALE24_TO_FLOAT and FROM_FLOAT macros which were only
used for SPIFloat and can easily be folded into the object itself.

Harmonize the default data type use with SPIEnum, SPILength and SPIFontSize
this allows the rendering engine nr-style.cpp to use the right rendering value
which is always the computed value and not need to think about the internal
implementation of the style variable between `value`, `computed` and will allow
these variable to be made private in the future.

All uses are explicit to rule out mistaken variable type use.

+++ src/gradient-chemistry.cpp
@@ -1784,7 +1784,7 @@ SPGradient *sp_gradient_vector_for_object( SPDocument *const doc, SPDesktop *con
             }
         } else if (paint.isColor()) {
             color = paint.getColor();
-            opacity = for_fill ? style.fill_opacity : style.stroke_opacity;
+            opacity = for_fill ? (double)style.fill_opacity : (double)style.stroke_opacity;
         }
     }
